+++ src/pages/Shared.js
@@ -4,8 +4,8 @@ import FooterElement from "./../components/common/FooterElement";
 import SharedSection from "./../components/shared/SharedSection";
 import Input from "../components/common/Input";
 import FolderList from "../components/common/FolderList";
-import useGetJson from "./../hook/uesGetJson";
-import "../style/shared.css";
+import useGetJson from "./../hooks/uesGetJson";
+import "../styles/shared.css";
 
 function Shared() {
   const foldersData = useGetJson(getFolderInfo);
